[리팩토링] 선호도 반환 DTO percent, count 타입 수정
int -> Long (count가 커지는 것에 대한 대비)
double -> Double

+++ src/main/java/com/cloth/wardrobe/dto/statistics/ContentForStatistics.java
@@ -9,6 +9,6 @@ import lombok.Data;
 public class ContentForStatistics {
     String name;
     StatisticsType type;
-    double percent;
-    int count;
+    Double percent;
+    Long count;
 }
